pyproject.toml
[project]
name = "kaspa-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]>=20.0",
    "websockets>=12.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "kaspa>=1.0.0",
    "grpcio>=1.75.0",
    "grpcio-tools>=1.75.0",
    "aiohttp>=3.12.15",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.1",
    "black>=24.10.0",
    "isort>=5.13.2",
]

[tool.black]
line-length = 100
target-version = ['py312']
exclude = '''
/(
    \.git
  | \.venv
  | src/protos
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_glob = ["src/protos/*.py"]
